+++ examples/resources/pingfederate_session_application_session_policies/import.sh
@@ -1,2 +1,0 @@
-# "sessionApplicationSessionPolicyId" should be "id"
-terraform import pingfederate_session_authentication_session_policies_global.mySessionApplicationSessionPolicy sessionApplicationSessionPolicyId